Split SvgQuadraticCurveSegment
